+++ packages/fcl/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - YYYY-MM-DD **BREAKING?** -- description
 
+### 0.0.67-alpha.[36..39] - 2020-12-03
+
+- 2020-12-03 -- Updates `@onflow/send` -- This includes some fixes involving how signatures are added to transaction rpc calls
+
 ### 0.0.67-alpha.[19..35] - 2020-11-26
 
 - 2020-11-26 -- `pre-authz` golden path tested on testnet with dapper