+++ packages/sdk-resolve/CHANGELOG.md
@@ -2,10 +2,14 @@
 
 - YYYY-MM-DD **BREAKING?** -- description
 
-### 0.0.4
+### 0.0.[5..9] -- 2020-12-03
+
+- 2020-12-03 -- VSN `@onflow/sdk-resolve-ref-block-id` 0.0.6
+
+### 0.0.4 -- 2020-11-04
 
 - 2020-11-04 -- VSN `@onflow/sdk-resolve-signatures` 0.0.4 -> 0.0.5
- 
+
 ### 0.0.3
 
 - 2020-10-28 -- VSN `@onflow/sdk-resolve-signatures` 0.0.3 -> 0.0.4